requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cmake==3.12.0
constantly==15.1.0
cryptography==2.3.1
cycler==0.10.0
Cython==0.28.5
dateparser==0.7.0
decorator==4.3.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
imutils==0.5.1
incremental==17.5.0
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.13.1
Jinja2==2.10
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
lmdb==0.94
lxml==4.2.5
MarkupSafe==1.0
matplotlib==3.0.0
mkl-fft==1.0.4
mkl-random==1.0.1
numpy==1.15.1
olefile==0.46
opencv-python==3.4.3.18
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==4.2.1
prompt-toolkit==2.0.5
psycopg2==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.4.2
pyOpenSSL==18.0.0
pyparsing==2.2.2
python-binance==0.7.0
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
regex==2018.8.29
requests==2.19.1
scipy==1.1.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
torch==0.4.1
torchvision==0.2.1
torchwordemb==0.0.8
tornado==5.1.1
traitlets==4.3.2
Twisted==18.7.0
txaio==18.8.1
tzlocal==1.5.1
urllib3==1.23
virtualenv==16.0.0
waitress==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
zope.interface==4.5.0
